requirements.txt
beautifulsoup4
dateparser
Gunicorn
Flask
requests
psycopg2-binary
flask_sqlalchemy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
google-genai
google-auth
flask-migrate
pymongo
